requirements.txt
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.7.2
auditlog3==1.0.1
backports.zoneinfo==0.2.1
billiard==4.1.0
bleach==3.3.0
boto3==1.26.32
botocore==1.29.165
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
configparser==5.0.2
cron-descriptor==1.4.0
cryptography==3.3.2
defusedxml==0.7.1
Deprecated==1.2.14
Django==3.2.3
django-advanced-filters==1.3.0
django-allow-cidr==0.6.0
django-appconf==1.0.5
django-braces==1.14.0
django-celery-beat==2.2.0
django-compressor==2.4.1
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-ebhealthcheck==2.0.2
django-extensions==3.1.3
django-imagekit==4.0.2
django-libsass==0.8
django-oauth-toolkit==1.7.1
django-oauth2-provider==0.2.6.1
django-ses==3.3.0
django-storages==1.13.1
django-timezone-field==4.2.3
djangorestframework==3.12.4
exceptiongroup==1.1.3
fonttools==4.40.0
fpdf==1.7.2
fpdf2==2.6.1
html5lib==1.1
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
jwcrypto==1.4.2
kombu==5.3.1
libsass==0.22.0
lml==0.1.0
mollie-api-python==2.5.3
oauthlib==3.2.2
olefile==0.46
packaging==23.1
pilkit==2.0
Pillow==10.0.0
# pkg_resources==0.0.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.5
pycparser==2.21
pyexcel==0.6.6
pyexcel-io==0.6.6
pyexcel-xls==0.6.2
PyPDF2==3.0.1
pypng==0.0.20
PyQRCode==1.2.1
python-bidi==0.4.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
rcssmin==1.0.6
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.1.0
s3transfer==0.6.1
sentry-sdk==1.4.3
shortuuid==1.0.11
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
texttable==1.6.7
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.5.0
wrapt==1.15.0
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0